pyproject.toml
[tool.poetry]
name = "softdesk-support-"
version = "0.1.0"
description = ""
authors = ["SiRipo92 <ripochesierra@gmail.com>"]
readme = "README.md"
packages = [{include = "softdesk_support"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-nested-routers = "^0.95.0"
drf-spectacular = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
ruff = "^0.14.10"
pip-audit = "^2.10.0"
django-extensions = "^4.1"
coverage = "^7.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
  ".venv",
  "**/migrations/**",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]


[tool.ruff.lint.isort]
known-first-party = ["apps", "common", "config"]
